net/mac80211/rc80211_minstrel_debugfs.c: fix return type
this warning:

  net/mac80211/rc80211_minstrel_debugfs.c:145: warning: initialization from incompatible pointer type

triggers because the proper return type for file_operations::read
is ssize_t, not int.

+++ net/mac80211/rc80211_minstrel_debugfs.c
@@ -106,7 +106,7 @@ minstrel_stats_open(struct inode *inode, struct file *file)
 	return 0;
 }
 
-static int
+static ssize_t
 minstrel_stats_read(struct file *file, char __user *buf, size_t len, loff_t *o)
 {
 	struct minstrel_stats_info *ms;
